Change "npm run dev" to postAttachCommand
Note to future self: nohup <command> & sends the command to the
background, and the process will not be killed when the terminal is
closed. Output is sent to nohup.out (added to .gitignore). If I just
used nohup, the command wouldn't be sent to the background but would in
theory live on when terminal is killed (in practice I don't think that
happens with codespaces). If I just used & then the command would be
sent to the background but the job would possibly be killed when the
terminal is exited and the output is just written to standard out. Using
both prevents the output from being cluttered up, maintains the output,
and is resilient to the terminal being closed.